audio: base_fw: move IPC4_EXTENDED_SYSTEM_TIME to platform code
Move implementation of IPC4_EXTENDED_SYSTEM_TIME from generic
to platform specific code. The implementation depends on specific
ART counter that is not available on all platforms, so it is better
to handle this in platform specific code.
